% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewtParsToMoments.R
\name{skewtParsToMoments}
\alias{skewtParsToMoments}
\title{Skew-t parameter transfornation}
\usage{
skewtParsToMoments(x)
}
\description{
Compute the first four moments from the four skew-t parameters
}
