% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stInfoMat.R
\name{stInfoMat}
\alias{stInfoMat}
\title{Skew-t Information Matrix Computation}
\usage{
stInfoMat(y = NULL, dp, type = c("expected", "observed"))
}
\description{
Compute the skew-t observed information analytically, and
compute expected information matrix numerically.
}
